broke db out into different files
