+++ trackscape-discord-bot/src/commands/toggle_broadcasts.rs
@@ -4,7 +4,7 @@ use serenity::client::Context;
 use serenity::model::prelude::application_command::{CommandDataOption, CommandDataOptionValue};
 use serenity::model::prelude::command::CommandOptionType;
 use serenity::model::prelude::Permissions;
-use trackscape_discord_shared::database::RegisteredGuildModel;
+use trackscape_discord_shared::database::guilds_db::RegisteredGuildModel;
 use trackscape_discord_shared::osrs_broadcast_extractor::osrs_broadcast_extractor::BroadcastType;
 use trackscape_discord_shared::osrs_broadcast_extractor::osrs_broadcast_extractor::BroadcastType::ItemDrop;
